% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AnnoyBuildObj}
\alias{AnnoyBuildObj}
\title{Build an Annoy object}
\usage{
AnnoyBuildObj(data.use, n.trees = 50)
}
\arguments{
\item{data.use}{Data used to create nearest-neighbor graph.}

\item{n.trees}{Number of trees used in Annoy search query. Higher value gives higher precision}
}
\value{
Returns an annoy object, as documented by RcppAnnoy
}
\description{
Build an Annoy object
}
