% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AnnoyGetNN}
\alias{AnnoyGetNN}
\title{Find nearest-neighbors of elements in an Annoy Object}
\usage{
AnnoyGetNN(
  obj,
  query,
  k.param = 30,
  search.k = -1,
  get.distance = TRUE,
  mc.cores = 1
)
}
\arguments{
\item{obj}{Annoy obj used to search for nearest-neighbors}

\item{query}{The data used to create the Annoy object, containing elements you want to find nearest-
neighbors of}

\item{k.param}{Number of nearest-neighbors to calculate.}

\item{search.k}{Number of nodes an Annoy search query touches. -1 invokes n.trees * n search}

\item{get.distance}{Include distances}

\item{mc.cores}{Number of cores to use in a parallel format}
}
\value{
Returns a list containing the nearest-neighbors of each element and the distances between
elements and their nearest-neighbors
}
\description{
Given an Annoy object, this function quickly finds the nearest-neighbors of each element
}
