% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BuildSNN}
\alias{BuildSNN}
\title{Build a shared nearest-neighbor graph}
\usage{
BuildSNN(
  data.use,
  k.param = 30,
  prune.SNN = 1/15,
  nn.eps = 0,
  method = "annoy",
  n.trees = 50,
  search.k = -1,
  mc.cores = 1
)
}
\arguments{
\item{data.use}{Data used to create nearest-neighbor graph.}

\item{k.param}{Number of nearest-neighbors to calculate.}

\item{prune.SNN}{Cutoff for removing edges signaling low shared overlap of neighbors.}

\item{nn.eps}{Error bound for using RANN to calculate nearest-neighbors.}

\item{method}{Method to use for creating nearest-neighbor graph. Annoy tends to be faster, but does not give
exact solution}

\item{n.trees}{Number of trees used in Annoy search query. Higher value gives higher precision}

\item{search.k}{Number of nodes an Annoy search query touches. -1 invokes n.trees * n search}

\item{mc.cores}{Number of cores to use in a parallel format}
}
\value{
Returns an snn object containing the shared nearest-neighborr graph
}
\description{
Returns a shared nearest-neighbor graph by first creating a nearest-neighbor
graph with Annoy and then compute shared nearest-neighbor with Seurat's cpp implementation
}
