% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcBinomCI}
\alias{calcBinomCI}
\title{Calculate a binomial confidence interval}
\usage{
calcBinomCI(p, n, z = 1.96)
}
\arguments{
\item{p}{The observed proportion}

\item{n}{The total number of samples/trials used to calculate the proportion}

\item{z}{The quantile of a standard normal distribution}
}
\value{
A singular numeric value representing the half-width of the CI
}
\description{
Given an observed proportion (p) and a number of trials (n), calculate the
half-width of a (z)\% binomial confidence interval
}
