% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createParamTable}
\alias{createParamTable}
\title{Create a table of parameters that will be used for simulations}
\usage{
createParamTable(
  nreps,
  clus,
  fc,
  ncases = 10,
  nctrls = 10,
  nbatches = 4,
  b_scale = 1,
  s_scale = 1,
  cf_scale = 1,
  res_use = 1.2,
  cond_induce = "cases",
  save_path
)
}
\arguments{
\item{nreps}{Numeric value indicating the number of replicates of a simulation that will be run.}

\item{clus}{The name of the cluster in which a fold change will be induced.}

\item{fc}{The magnitude of the fold change that will be induced in the chosen cluster. If no fold change is desired, set
fc = 1.}

\item{ncases}{The number of cases.}

\item{nctrls}{The number of controls.}

\item{nbatches}{The number of batches that samples will be distributed into.}

\item{b_scale}{The magnitude of batch-associated gene expression variation the simulated dataset will exhibit. Setting b_scale
= 1 will result in realistic levels of batch-associated variation (as derived from parameter estimation of the input dataset).
Increasing b_scale results in higher variation, while decreasing b_scale results in lower variation.}

\item{s_scale}{The magnitude of sample-associated gene expression variation the simulated dataset will exhibit. Setting s_scale
= 1 will result in realistic levels of sample-associated variation (as derived from parameter estimation of the input dataset).
Increasing s_scale results in higher variation, while decreasing s_scale results in lower variation.}

\item{cf_scale}{The magnitude of cell state frequency variation that cell states will exhibit across samples in the simulated
dataset. Setting cf_scale = 1 will result in realistic levels of cell state frequency variation (as derived from parameter
estimation of the input dataset). Increasing cf_scale results in higher variation, while decreasing cf_scale results in lower variation.}

\item{res_use}{The resolution that will be used for clustering (Louvain method) the simulated dataset.}

\item{cond_induce}{The condition you wish to induce a fold change in. Setting cond_induce = "cases" will induce a fold
change into cases, while setting cond_induce = "ctrls" will induce a fold change into controls.}

\item{save_path}{The name of the directory the results will be saved to.}
}
\value{
A data.frame containing user-controlled parameters that will be used for simulations
}
\description{
Create a table of parameters that will be used for simulations
}
