% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributeSampleFuncs.R
\name{distribSamplePerBatch}
\alias{distribSamplePerBatch}
\title{Distribute each sample into its own batch}
\usage{
distribSamplePerBatch(ncases, nctrls, nbatches)
}
\arguments{
\item{ncases}{The number of cases.}

\item{nctrls}{The number of controls.}

\item{nbatches}{The number of batches that samples will be distributed into.}
}
\value{
Returns a list containing the following elements: a list containing the batch
structure of the study (info on which samples are cases/controls and which batch they
were placed in), a vector containing the names of all samples, a vector containing the
names of case samples, and a vector containing the names of control samples. This output is used directly
as input into our simulation functions (batchStructure parameter).
}
\description{
Given the number of cases, the number of controls, and number of batches, this function will
distribute a sample into its own batch. The number of bathes MUST be equal to the total
number of samples (nbatch = ncases + nctrls)
}
