% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributeSampleFuncs.R
\name{distribSampleSplit}
\alias{distribSampleSplit}
\title{Split each sample into a specified number of subsamples before distributing into batches}
\usage{
distribSampleSplit(ncases, nctrls, nbatches, numSubsamples)
}
\arguments{
\item{ncases}{The number of cases.}

\item{nctrls}{The number of controls.}

\item{nbatches}{The number of batches that samples will be distributed into.}

\item{numSubsamples}{The number of subsamples that each sample will be divided into}
}
\value{
Returns a list containing the following elements: a list containing the batch
structure of the study (info on which samples are cases/controls and which batch they
were placed in), a vector containing the names of all samples, a vector containing the
names of case samples, and a vector containing the names of control samples. This output is used directly
as input into our simulation functions (batchStructure parameter).
}
\description{
Given the number of cases, the number of controls, and number of batches, this function will
first split each sample (regardless of case-control status) into a specified number of
equally sized subsamples (numSubsamples). These subsamples will then be distributed into
different batches. Each batch will contain cells from multiple samples (equal to numSubsamples)
}
