% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributeSampleFuncs.R
\name{distributeSamples}
\alias{distributeSamples}
\title{Distribute samples equally into batches in equal numbers of cases and controls}
\usage{
distributeSamples(ncases, nctrls, nbatches)
}
\arguments{
\item{ncases}{The number of cases.}

\item{nctrls}{The number of controls.}

\item{nbatches}{The number of batches that samples will be distributed into.}
}
\value{
Returns a list containing the following elements: a list containing the batch
structure of the study (info on which samples are cases/controls and which batch they
were placed in), a vector containing the names of all samples, a vector containing the
names of case samples, and a vector containing the names of control samples. This output is used directly
as input into our simulation functions (batchStructure parameter).
}
\description{
Given the number of cases, number of controls, and number of batches in a planned study design
this function distributes samples equally into batches, in equal numbers of cases and controls
based off the "split" function.
}
