% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterEstimation.R
\name{estimateFreqVar}
\alias{estimateFreqVar}
\title{Estimate the mean and variation in each cell state's frequency across samples}
\usage{
estimateFreqVar(meta, clusCol, sampleCol, logCov = TRUE)
}
\arguments{
\item{meta}{A metadata table that should include named columns containing: Cell IDs,
cell state assignments, and sample identities for each cell.}

\item{clusCol}{The name of the metadata column containing the cell state cluster assignments for each cell.}

\item{sampleCol}{The name of the metadata column containing the sample identities of each cell.}

\item{logCov}{Boolean designating whether you want the variance-covariance matrix in log space (default)
or in linear space. The log space matrix is used as input to our simulation functions; only use
linear space for visualization.}
}
\value{
Returns a list containing: a vector containing the mean frequency
of each cell state (in linear space) and a variance-covariance matrix containing the covariances
between each cell state (log space by default). The mean frequency vector will be transformed into
log space during data simulation.
}
\description{
Given a metadata table containing each cell's sample identity and assigned cell
state, this function will return a list containing: a vector containing the mean frequency
of each cell state (in linear space) and a variance-covariance matrix containing the covariances
between each cell state (log space by default). Both of these elements are used as direct
inputs to our dataset generation functions.
}
