% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePCs.R, R/powerCalcFuncs.R,
%   R/simulateDatasets.R
\name{generateSampleShifts}
\alias{generateSampleShifts}
\title{Simulate sample shifts for each cell state}
\usage{
generateSampleShifts(sample_list, sample_vars, s_scale)

generateSampleShifts(sample_list, sample_vars, s_scale)

generateSampleShifts(sample_list, sample_vars, s_scale)
}
\arguments{
\item{sample_list}{A named list of the sample names e.g. sample1, sample2, etc.}

\item{sample_vars}{A matrix containing the sample_associated variance for each cell state in each PC. These are obtained
as output from the "estimatePCVar" function.}

\item{s_scale}{The magnitude of sample-associated gene expression variation the simulated dataset will exhibit. Setting s_scale
= 1 will result in realistic levels of sample-associated variation (as derived from parameter estimation of the input dataset).
Increasing s_scale results in higher variation, while decreasing s_scale results in lower variation.}
}
\value{
Returns a matrix containing the sample shift for each cell state in each PC

Returns a matrix containing the sample shift for each cell state in each PC

Returns a matrix containing the sample shift for each cell state in each PC
}
\description{
This function generates linear shifts in each samplefor each cell state. These shifts contribute to the adjusted centroid
that is used for simulating PC locations

This function generates linear shifts in each samplefor each cell state. These shifts contribute to the adjusted centroid
that is used for simulating PC locations

This function generates linear shifts in each samplefor each cell state. These shifts contribute to the adjusted centroid
that is used for simulating PC locations
}
