% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePCs.R, R/powerCalcFuncs.R,
%   R/simulateDatasets.R
\name{generateSamplesInBatches}
\alias{generateSamplesInBatches}
\title{Simulate the samples in a single designated batch}
\usage{
generateSamplesInBatches(
  batchStructure,
  batch = "batch1",
  case_freqs,
  ctrl_freqs,
  cfcov,
  cf_scale,
  centroids,
  batch_shifts,
  sample_shifts,
  pc_cov_list,
  ncells
)

generateSamplesInBatches(
  batchStructure,
  batch = "batch1",
  case_freqs,
  ctrl_freqs,
  cfcov,
  cf_scale,
  centroids,
  batch_shifts,
  sample_shifts,
  pc_cov_list,
  ncells
)

generateSamplesInBatches(
  batchStructure,
  batch = "batch1",
  case_freqs,
  ctrl_freqs,
  cfcov,
  cf_scale,
  centroids,
  batch_shifts,
  sample_shifts,
  pc_cov_list,
  ncells
)
}
\arguments{
\item{batchStructure}{The structure of the study design in which cases and controls are split into batches. These
structures are output by the "distributeSample" functions (which can then be modified if specific structure is desired).
If this parameter is kept as NULL, this function will automatically create a batchStructure with the "distributeSamples"
function.}

\item{batch}{The name of the batch that will be simulated}

\item{case_freqs}{The cell state frequency distributions for all case samples. These are received as output from the
"generateFreqs" function}

\item{ctrl_freqs}{The cell state frequency distributions for all control samples. These are received as output from the
"generateFreqs" function}

\item{cfcov}{The cell state frequency variance-covariance across samples. This is obtained as output from the "estimateFreqVar"
function}

\item{cf_scale}{The magnitude of cell state frequency variation that cell states will exhibit across samples in the simulated
dataset. Setting cf_scale = 1 will result in realistic levels of cell state frequency variation (as derived from parameter
estimation of the input dataset). Increasing cf_scale results in higher variation, while decreasing cf_scale results in lower variation.}

\item{centroids}{The mean PC values for each cell state. These are obtained as output from the "estimatePCVar" function.}

\item{batch_shifts}{The simulated linear shifts in each batch for each cell state. These are received as output from the
"generateBatchShifts" function.}

\item{sample_shifts}{The simulated linear shifts in each sample for each cell state. These are received as output from the
"generateSampleShifts" function.}

\item{pc_cov_list}{A list containing the residual variance-covariance matrices for each cell state. These are obtained
as output from the "estimatePCVar" function.}

\item{ncells}{A vector containing the number of cells that will be simulated for each sample. The vector must be
the same length as the number of total samples (ncases + nctrls).}
}
\value{
Returns a metadata table and the simulated PC locations for all simulated cells in the designated batch

Returns a metadata table and the simulated PC locations for all simulated cells in the designated batch

Returns a metadata table and the simulated PC locations for all simulated cells in the designated batch
}
\description{
This function generates the principal component (PC) location for all simulated cells in a single designated
batch.

This function generates the principal component (PC) location for all simulated cells in a single designated
batch.

This function generates the principal component (PC) location for all simulated cells in a single designated
batch.
}
