% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFrequencies.R
\name{getInducedFreqs}
\alias{getInducedFreqs}
\title{Generate a cell state frequency distribution with induced frequencies}
\usage{
getInducedFreqs(log_prior, clus, fc = 1, cf_sigma)
}
\arguments{
\item{log_prior}{A named vector containing the mean frequencies of the prototype dataset's cell states (log space). The
"estimateFreqVar" function returns this a mean frequency vector in linear space (can be transformed into log space via the
"log" function).}

\item{clus}{The name of the cluster in which a fold change will be induced.}

\item{fc}{The magnitude of the fold change that will be induced in the chosen cluster. If no fold change is desired, set
fc = 1.}

\item{cf_sigma}{A matrix containing the covariance between cell states. This matrix is received as output from
the "estimateFreqVar" function}
}
\value{
Returns a generated vector in which the designated cluster has the designated fold change induced
}
\description{
Given a vector of cell state frequencies (log space), this function will induce a specified fold change (fc)
into the specified cluster (clus)
}
