% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterEstimation.R
\name{getModVars}
\alias{getModVars}
\title{Retrieve variables from models fitted in the "estimatePCVar" function}
\usage{
getModVars(mods, npcs, clusNames)
}
\arguments{
\item{mods}{A list of lme4 odels that this function will extract the variables from}

\item{npcs}{The number of PCs whose variance will be deconvoluted. As this function's output is used
as direct input for our simulation functions, this will determine how many PCs your generated datasets
will contain (e.g. npcs = 20 here will mean that the generated datasets will contain 20 PCs).}

\item{clusNames}{A list of unique names of the cell state cluster assignments}
}
\value{
Returns a list containing the following elements: a matrix containing the mean PC values for each
cell state, a list where each element is the residual variance-covariance matrix for each cell state,
a matrix containing the batch-associated variance for each cell state in each PC, a matrix containing the
sample-associated variance for each cell state in each PC.
}
\description{
Given fitted models and the names for each cell state, retrieve for each cell state: the centroids
(fixed effect), residual variance-covariance matrices (residuals), batch-associated variances (random effect variance),
and sample associated variances (random effect variance).
}
