% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getParsFromResFile}
\alias{getParsFromResFile}
\title{Retrieve the parameters from a simulation}
\usage{
getParsFromResFile(fileList)
}
\arguments{
\item{fileList}{A list of filenames that refer to the results from simulating a dataset}
}
\value{
Returns a dataframe containing the parameters used to simulate a dataset
}
\description{
Given a list of files that contain the results from simulating a dataset (via simDataset.base or simDataset.withMASC),
this function will retrieve the parameters used to simulate the dataset
}
