% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPowerFromRes}
\alias{getPowerFromRes}
\title{Retrieve the power results from simulations}
\usage{
getPowerFromRes(
  resFiles,
  resTables,
  threshold = 0.05,
  z = 1.96,
  stratByClus = FALSE
)
}
\arguments{
\item{resFiles}{A list of filenames that refer to the results from simulating a dataset}

\item{resTables}{The MASC results tables obtained from MASC analysis on simulated datasets}

\item{threshold}{The p-value threshold that determines significance}

\item{z}{Determines the size of the confidence interval (CI) where z represents the \% CI}

\item{stratByClus}{Boolean that determines whether the power results will be calculated as the aggregate of all
clusters or stratified by cluster.}
}
\value{
Returns a dataframe containing the power calculations from the simulated datasets
}
\description{
Given results, this function will calculate the number of simulations in which at least one cell state
cluster had a significant p-value frorm MASC analysis, meaning the cluster was expanded or depleted between
conditions.
}
