% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerCalcFuncs.R
\name{getPvals.MASC}
\alias{getPvals.MASC}
\title{Test clusters for an expansion (or depletion) between conditions}
\usage{
getPvals.MASC(
  meta,
  clusterCol,
  null_mod,
  full_mod,
  mc.cores = 1,
  adj_method = "bonferroni"
)
}
\arguments{
\item{meta}{A metadata table received as output from simulating datasets (via simDataset.base or simDataset.withMASC). The
metadata table should contain the following columns: "cellstate" which refers to the assigned cell state during simulation,
"sample" which refers to the sample the cell was assigned to, "batch" which refers to the batch the cell was assigned to, and
"condition" which refers to the condition the cell was assigned to (case or control).}

\item{clusterCol}{The name of the metadata column containing the new cell state clusters. Our functions designate this column
"new_clus" in the saved results.}

\item{null_mod}{The right-hand side of the formula that will be used as the null model in MASC analysis}

\item{full_mod}{The right-hand side of the formula that will be used as the full model in MASC analysis}

\item{mc.cores}{The number of cores that will be used for simulating the dataset and clustering. The computation time will
depend on the number of simulated cells.}

\item{adj_method}{The p-value correction method that will be used via the "p.adjust" function}
}
\value{
Returns a dataframe containing the calculated MASC p-values for each new cluster, as well as
corrected p-values (default Bonferroni corrected)
}
\description{
Given the results of a simulated dataset (meta), this function will test all of the new clusters
for differential abundance (expansion or depletion) between conditions.
}
