% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFrequencies.R
\name{induceFC.capped}
\alias{induceFC.capped}
\title{Induce a fold change in an element in a cluster}
\usage{
induceFC.capped(baseline, clus, fc = 1)
}
\arguments{
\item{baseline}{A named vector containing the frequencies of cell states}

\item{clus}{The name of the cluster in which a fold change will be induced.}

\item{fc}{The magnitude of the fold change that will be induced in the chosen cluster. If no fold change is desired, set
fc = 1.}
}
\value{
Returns a vector in which the designated cluster has the designated fold change induced
}
\description{
Given a baseline distribution, induce a fold change in a cluster
}
