% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insertElems}
\alias{insertElems}
\title{Insert an element or vector of elements to designated positions in a vector}
\usage{
insertElems(vect, pos, elems)
}
\arguments{
\item{vect}{The vector you are inserting elements into}

\item{pos}{The position you want to insert an element/s into}

\item{elems}{The element/s you want to insert into a vector}
}
\value{
A vector containing both the original and inserted elements
}
\description{
Implementation from Tutur Qhuhuit that inserts an element/s into a vector
}
