% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerCalcFuncs.R
\name{mascCalc}
\alias{mascCalc}
\title{Perform a test for differential abundance for a cell state cluster with MASC}
\usage{
mascCalc(meta, clusterCol, clusterName, null_mod, full_mod)
}
\arguments{
\item{meta}{A metadata table received as output from simulating datasets (via simDataset.base or simDataset.withMASC). The
metadata table should contain the following columns: "cellstate" which refers to the assigned cell state during simulation,
"sample" which refers to the sample the cell was assigned to, "batch" which refers to the batch the cell was assigned to, and
"condition" which refers to the condition the cell was assigned to (case or control).}

\item{clusterCol}{The name of the metadata column containing the new cell state clusters. Our functions designate this column
"new_clus" in the saved results.}

\item{clusterName}{The name of the cell state cluster that will be tested forr differential abundance.}

\item{null_mod}{The right-hand side of the formula that will be used as the null model in MASC analysis}

\item{full_mod}{The right-hand side of the formula that will be used as the full model in MASC analysis}
}
\value{
Returns a p-value (numeric) obtained from MASC analysis

importFrom dplyr group_by tally pull
importFrom lme4 glmer glmerControl .makeCC
}
\description{
Given a metadata table, this function will test a desginated cell state cluster (clusterName) for differential abundance (expansion
or depletion between conditions)
with MASC
}
