% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ra_FibObj}
\alias{ra_FibObj}
\title{List of metadata and PC embeddings for only the fibroblasts from the RA dataset before running Harmony integration.
Thie was used for Fig 4a}
\format{
meta: data.frame of 5,265 rows (cells) and 10 columns. "clusAllCells" refers to
cluster assignments from clustering all of the RA data, while "clusOnlyFib"
refers to cluster assignments from clustering only the fibroblasts (this was used in Fig 4a)
embeddings: data.frame of 5,265 row (cells) and 20 columns (PCs).
}
\source{
\url{https://www.immport.org/shared/study/SDY998}
}
\usage{
ra_FibObj
}
\description{
List of metadata and PC embeddings for only the fibroblasts from the RA dataset before running Harmony integration.
Thie was used for Fig 4a
}
\keyword{datasets}
