% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ra_HarmObj}
\alias{ra_HarmObj}
\title{List of metadata and PC embeddings for the RA dataset after running Harmony integration}
\format{
meta: data.frame of 5,265 rows (cells) and 10 columns. "preHarmClus" refers to
cluster assignments from clustering the data before Harmony, while "harmClus"
refers to cluster assignments from clustering the data after Harmony
embeddings: data.frame of 5,265 row (cells) and 20 columns (harmonized PCs). Obtained from Harmony integration
}
\source{
\url{https://www.immport.org/shared/study/SDY998}
}
\usage{
ra_HarmObj
}
\description{
List of metadata and PC embeddings for the RA dataset after running Harmony integration
}
\keyword{datasets}
