% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpost_seurat_cluster.R
\name{scpost_seurat_cluster}
\alias{scpost_seurat_cluster}
\title{Function to perform Seurat clustering using a Seurat object generated with scpost_seurat_init}
\usage{
scpost_seurat_cluster(
  seurat.obj = seurat.obj,
  pc_max = 10,
  resolution = 0.05,
  reduction = "pca",
  seurat_return = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{seurat.obj}{A Seurat object generated with scpost_seurat_init}

\item{pc_max}{Number of PCs to utilize for clustering}

\item{resolution}{Numeric "resolution" value that dictates cluster size}

\item{reduction}{Default is "pca"; Select "harmony" if a batch_correction variable was specified on scpost_parameters.csv (batch_correction != FALSE)}

\item{seurat_return}{If TRUE, returns a Seurat object; if FALSE, returns cluster assignments}

\item{verbose}{TRUE or FALSE; print Seurat package messages}
}
\value{
Seurat object or cluster assignments
}
\description{
This function applies Seurat package functions to a Seurat object in order to find nearest neighbors & perform SNN & louvain community detection
}
\examples{
scpost_seurat_cluster(seurat.obj=seurat.obj, npc=10, resolution=0.05, reduction="pca", seurat_return=FALSE, verbose=TRUE)
}
