% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpost_seurat_init.R
\name{scpost_seurat_init}
\alias{scpost_seurat_init}
\title{Function to build a Seurat object, then normalize, scale & run PCA}
\usage{
scpost_seurat_init(counts = counts, pc_max = 10, verbose = TRUE)
}
\arguments{
\item{counts}{Counts matrix}

\item{pc_max}{Max number of PCs to incorporate from PCA}

\item{verbose}{TRUE or FALSE; print Seurat package messages}
}
\value{
Seurat object
}
\description{
This function aggregates several Seurat package functions to build a Seurat object, then normalize data, scale data, & run PCA
}
\examples{
scpost_seurat_init(counts=counts, pc_max=10, verbose=TRUE)
}
