% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDatasets.R
\name{simDataset.base}
\alias{simDataset.base}
\title{Simulate a dataset and save the results}
\usage{
simDataset.base(
  save_path,
  rep = 1,
  seed = 1,
  ncases,
  nctrls,
  nbatches,
  batchStructure = NULL,
  ncells,
  centroids,
  pc_cov_list,
  batch_vars,
  b_scale = 1,
  sample_vars,
  s_scale = 1,
  cfcov,
  cf_scale = 1,
  meanFreqs,
  clus,
  fc = 1,
  cond_induce = "cases",
  res_use = 1.2,
  mc.cores = 1,
  clusterData = TRUE,
  returnPCs = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{save_path}{The name of the directory the results will be saved to.}

\item{rep}{A numeric value representing the replicate number of the simulated dataset.}

\item{seed}{A numeric value representing the seed that will be set before simulating the dataset.}

\item{ncases}{The number of cases.}

\item{nctrls}{The number of controls.}

\item{nbatches}{The number of batches that samples will be distributed into.}

\item{batchStructure}{The structure of the study design in which cases and controls are split into batches. These
structures are output by the "distributeSample" functions (which can then be modified if specific structure is desired).
If this parameter is kept as NULL, this function will automatically create a batchStructure with the "distributeSamples"
function.}

\item{ncells}{A vector containing the number of cells that will be simulated for each sample. The vector must be
the same length as the number of total samples (ncases + nctrls).}

\item{centroids}{The mean PC values for each cell state. These are obtained as output from the "estimatePCVar" function.}

\item{pc_cov_list}{A list containing the residual variance-covariance matrices for each cell state. These are obtained
as output from the "estimatePCVar" function.}

\item{batch_vars}{A matrix containing the batch_associated variance for each cell state in each PC. These are obtained
as output from the "estimatePCVar" function.}

\item{b_scale}{The magnitude of batch-associated gene expression variation the simulated dataset will exhibit. Setting b_scale
= 1 will result in realistic levels of batch-associated variation (as derived from parameter estimation of the input dataset).
Increasing b_scale results in higher variation, while decreasing b_scale results in lower variation.}

\item{sample_vars}{A matrix containing the sample_associated variance for each cell state in each PC. These are obtained
as output from the "estimatePCVar" function.}

\item{s_scale}{The magnitude of sample-associated gene expression variation the simulated dataset will exhibit. Setting s_scale
= 1 will result in realistic levels of sample-associated variation (as derived from parameter estimation of the input dataset).
Increasing s_scale results in higher variation, while decreasing s_scale results in lower variation.}

\item{cfcov}{The cell state frequency variance-covariance across samples. This is obtained as output from the "estimateFreqVar"
function}

\item{cf_scale}{The magnitude of cell state frequency variation that cell states will exhibit across samples in the simulated
dataset. Setting cf_scale = 1 will result in realistic levels of cell state frequency variation (as derived from parameter
estimation of the input dataset). Increasing cf_scale results in higher variation, while decreasing cf_scale results in lower variation.}

\item{meanFreqs}{A vector containing the mean frequencies of cell states (linear space) across samples from the original input
dataset. This vector is obtained as output from the "estimateFreqVar" function.}

\item{clus}{The name of the cluster in which a fold change will be induced.}

\item{fc}{The magnitude of the fold change that will be induced in the chosen cluster. If no fold change is desired, set
fc = 1.}

\item{cond_induce}{The condition you wish to induce a fold change in. Setting cond_induce = "cases" will induce a fold
change into cases, while setting cond_induce = "ctrls" will induce a fold change into controls.}

\item{res_use}{The resolution that will be used for clustering (Louvain method) the simulated dataset.}

\item{mc.cores}{The number of cores that will be used for simulating the dataset and clustering.}

\item{clusterData}{Boolean determining whether the simulated dataset will be clustered.}

\item{returnPCs}{Boolean determining whether the function will also return the simulated PC locations for reach cell.}

\item{verbose}{Print out time dataset was simulated at}
}
\value{
This function returns NULL and instead saves the results to the directory designated in "save_path". If returnPCs = FALSE,
the saved results will be a list containing the metadata table for the simulated cells. The metadata table contains a dataframe with
the following columns: "cellstate" which refers to the assigned cell state during simulation, "sample" which refers to the sample the
cell was assigned to, and "condition" which refers to the condition the cell was assigned to (case or control). If clusterData = TRUE,
the metadata table will also contain a column "new_clus", which refers to the new cluster assignments for the simulated cells.
If returnPCs = TRUE, the saved results will contain the metadata table for the simulated cells and a matrix containing the simulated
PC locations for each cell.
}
\description{
This function simulates a dataset and then saves the results as a list containing a metadata table for all simulated cells. The simulated
PC locations for all cells can optionally be saved.
}
