% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_scpost.R
\name{template_scpost}
\alias{template_scpost}
\title{Template function for post-processing scRNA-Seq data}
\usage{
template_scpost(dir_output = dir_output)
}
\arguments{
\item{dir_output}{Output directory for post-processing analysis results}
}
\value{
ExpressionSet object
}
\description{
This function reads an ExpressionSet object with scRNA data and performs core functionalities for post-processing of scRNA-Seq data
}
\examples{
template_scpost(dir_output="/path/to/output")
}
