% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zero_range}
\alias{zero_range}
\title{Simple implementation from Hadley Wickham that checks if all elements
of a vector are equal}
\usage{
zero_range(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Vector containing numeric elements}

\item{tol}{The tolerated acceptable error in determining whether elements are equal}
}
\value{
A boolean representing if all elements of a vector are equal
}
\description{
Simple implementation from Hadley Wickham that checks if all elements
of a vector are equal
}
